transformation-load/requirements.txt
# Required API packages
python-dotenv==1.0.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.20
alembic==1.11.3
pandas==2.0.3
